requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
gunicorn==21.2.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
sendgrid==6.10.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
